Fixing Goods Method and many more🌟🍭.

+++ Ruay/src/main/java/com/ruaymak3/Ruay/dto/InvoiceDetailDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class InvoiceDetailDto {
     private Long id;            // รหัสรายละเอียดสินค้า
-    private Integer goodId;      // รหัสสินค้า
+    private Long goodId;      // รหัสสินค้า
     private Integer quantity;    // จำนวนสินค้า
     private Double amount;       // ราคาสินค้าต่อชิ้น
     private Integer status;
